dont case sizeof to int

+++ mount_wokfs/main.c
@@ -44,7 +44,7 @@ unmount(const char *mp)
 	    mp, mp, mp);
 	if (rc == -1)
 		psc_fatal("snprintf: umount %s", mp);
-	if (rc >= (int)sizeof(buf))
+	if ((size_t)rc >= sizeof(buf))
 		psc_fatalx("snprintf: umount %s: too long", mp);
 	if (system(buf) == -1)
 		psclog_warn("system(%s)", buf);
